lesson_12_deploy_smart_contract/requirements.txt
python-dotenv==1.0.1
loguru==0.7.2
web3==6.15.1
fake-useragent==1.5.0
aiohttp==3.9.3
aiohttp-socks==0.8.4
curl_cffi==0.6.2
py-solc-x==2.0.2
eth-tester==0.10.0b4
py-evm==0.10.0b4